requirements.txt
python-telegram-bot[job-queue]>=20.0
python-dotenv>=1.0.0
pytz>=2024.1
APScheduler>=3.6.3
pytest>=8.3.5
pytest-asyncio>=0.25.3
aiohttp>=3.8.0
prometheus-client>=0.16.0
